refactor: Console 관련 기능 패키지 생성
