+++ src/main/java/chess/ConsoleApplication.java
@@ -1,6 +1,8 @@
 package chess;
 
-public class Application {
+import chess.console.ChessGame;
+
+public class ConsoleApplication {
 
     public static void main(String[] args) {
         ChessGame chessGame = new ChessGame();
