+++ src/main/java/chess/console/gamestate/End.java
@@ -1,4 +1,4 @@
-package chess.gamestate;
+package chess.console.gamestate;
 
 public class End implements GameState {
