+++ src/main/java/chess/console/gamestate/GameState.java
@@ -1,4 +1,4 @@
-package chess.gamestate;
+package chess.console.gamestate;
 
 public interface GameState {
