+++ src/main/java/chess/console/gamestate/Running.java
@@ -1,8 +1,8 @@
-package chess.gamestate;
+package chess.console.gamestate;
 
+import chess.console.view.OutputView;
 import chess.domain.ChessBoard;
 import chess.domain.Color;
-import chess.view.OutputView;
 
 public abstract class Running implements GameState {
