+++ src/test/java/chess/console/gamestate/EndTest.java
@@ -1,4 +1,4 @@
-package chess.gamestate;
+package chess.console.gamestate;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
